--- app/profile/[userKey]/page.tsx
@@ -1,10 +1,0 @@
-import PlatformProfiles from "#app/profile/[userKey]/(platform)/PlatformProfiles.jsx";
-
-export default function Profile({ params }: { params: { userKey: string } }) {
-  return (
-    <>
-      <h1>프로필</h1>
-      <PlatformProfiles />
-    </>
-  );
-}
