+++ app/profile/page.tsx
@@ -1,0 +1,9 @@
+import PlatformProfiles from "#app/profile/(platform)/PlatformProfiles.jsx";
+
+export default function Profile() {
+  return (
+    <>
+      <PlatformProfiles />
+    </>
+  );
+}
